COMMERCE-618 Move Go files into a subdirectory
